Here is the next testimonial:
image: Evan.jpg

Head of Strategy

"Alex was one of the most significant members of the Aragon team throughout 2022 and 2023. He is extraordinarily proactive, doesn't wait to be told what to do, and will always give his ear to bounce ideas off of - now matter how esoteric they might be! If anyone needs this kind of dynamic energy on their team, it is an absolute no brainer that they should speak with Alex.

A little bit on our background working together - while he initially joined to help our teams with Agile, Scrum, etc., the reality was that we were in a small and constantly changing environment, so the needs of every contributor evolved as much. He took this as an opportunity to flexibly do a little bit of everything across the entire organization without getting comfortable doing just what he was hired for. When he eventually evolved into a more Human Resources-related role, he was a natural. (Note to HR professionals: there is also something nice about having someone with a background in Agile to better understand the culture of working rather than just culture in the abstract)

He also led the planning for several of our companywide offsites, and they were widely considered by attendees as some of the best ones we've had."